fix: servidor agora funciona

+++ src/main/java/br/uninter/medalerta/repository/AlertaRepository.java
@@ -3,5 +3,5 @@ package br.uninter.medalerta.repository;
 import br.uninter.medalerta.model.Alerta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlertaRepository extends JpaRepository<Alerta, Long> {
+public interface AlertaRepository extends JpaRepository<Alerta, Integer> {
 }
